refactor(cli): extract resolveVaultDbPath util, fix per-session error isolation, gitignore settings.local.json
