+++ packages/forge/src/scaffold-filetree.ts
@@ -627,6 +627,9 @@ See [Hook Packs documentation](https://soleri.dev/docs/guides/pack-authoring/) f
       'AGENTS.md',
       'instructions/_engine.md',
       '',
+      '# Local config — machine-specific, do not commit',
+      'settings.local.json',
+      '',
       '# OS',
       '.DS_Store',
       '',
